+++ src/services/index.js
@@ -1,7 +1,9 @@
 import { permission as permissionManager } from './permissions'
 import * as downloadManager from './downloaders'
+import * as toastManager from './toasters'
 
 export const service = {
   permissionManager,
-  downloadManager
+  downloadManager,
+  toastManager,
 }
